+++ lib/src/preconditions_impl.dart
@@ -14,6 +14,19 @@ Logger _log = Logger("Preconditions");
 ///
 typedef FutureOr<PreconditionStatus> PreconditionFunction();
 
+/// Unique identificator of precondition.
+class PreconditionId {
+  final dynamic _value;
+
+  PreconditionId(this._value);
+
+  @override
+  bool operator ==(Object other) => identical(this, other) || other is PreconditionId && runtimeType == other.runtimeType && _value == other._value;
+
+  @override
+  int get hashCode => _value.hashCode;
+}
+
 /// Optionally provide this Widget builder to render a feedback to your user, i.e. "Please grant all permissions", etc.
 /// Example:
 ///
@@ -48,7 +61,7 @@ enum DependenciesStrategy {
 ///
 ///
 class PreconditionsRepository extends ChangeNotifier {
-  final Map<String, Precondition> _known = {};
+  final Map<PreconditionId, Precondition> _known = {};
   int _runningCount = 0;
 
   /// Use this flag to render CircularProgressIndicator or similar feedback to user.
@@ -65,8 +78,8 @@ class PreconditionsRepository extends ChangeNotifier {
   /// * allow precondition to cache its positive and/or negative result with [satisfiedCache] and [notSatisfiedCache].
   /// * specify [dependsOn] - set of preconditions which must be satisfied before the repository attempts to evaluate this one
   ///
-  Precondition registerPrecondition(String id, PreconditionFunction preconditionFunction,
-      {Iterable<String> dependsOn: const [],
+  Precondition registerPrecondition(PreconditionId id, PreconditionFunction preconditionFunction,
+      {Iterable<PreconditionId> dependsOn: const [],
       resolveTimeout: const Duration(seconds: 10),
       satisfiedCache: Duration.zero,
       notSatisfiedCache: Duration.zero,
@@ -79,10 +92,7 @@ class PreconditionsRepository extends ChangeNotifier {
       throw Exception("Precondition '$id' is already registered");
     }
     var _p = Precondition._(id, preconditionFunction, statusBuilder ?? _nullBuilder, Set.unmodifiable(dependsOn), this,
-        satisfiedCache: satisfiedCache,
-        notSatisfiedCache: notSatisfiedCache,
-        resolveTimeout: resolveTimeout,
-        dependenciesStrategy: dependenciesStrategy);
+        satisfiedCache: satisfiedCache, notSatisfiedCache: notSatisfiedCache, resolveTimeout: resolveTimeout, dependenciesStrategy: dependenciesStrategy);
     _known[id] = _p;
     _log.info("Registering $_p");
     notifyListeners();
@@ -95,11 +105,8 @@ class PreconditionsRepository extends ChangeNotifier {
   ///
   /// Use this mechanism to organize your preconditions into groups with different priority or purpose.
   ///
-  Precondition registerAggregatePrecondition(String id, Iterable<String> dependsOn,
-      {resolveTimeout: const Duration(seconds: 10),
-      satisfiedCache: Duration.zero,
-      notSatisfiedCache: Duration.zero,
-      StatusBuilder? statusBuilder}) {
+  Precondition registerAggregatePrecondition(PreconditionId id, Iterable<PreconditionId> dependsOn,
+      {resolveTimeout: const Duration(seconds: 10), satisfiedCache: Duration.zero, notSatisfiedCache: Duration.zero, StatusBuilder? statusBuilder}) {
     return registerPrecondition(id, () => PreconditionStatus.satisfied(),
         dependsOn: dependsOn,
         resolveTimeout: resolveTimeout,
@@ -116,12 +123,12 @@ class PreconditionsRepository extends ChangeNotifier {
   /// The [registerPrecondition.satisfiedCache] and [registerPrecondition.notSatisfiedCache]
   /// allow usage of previously obtained result of evaluation.
   ///
-  Future<Iterable<Precondition>> evaluatePreconditions() async {
+  Future<Iterable<Precondition>> evaluatePreconditions({bool ignoreCache: false}) async {
     var list = _known.values.toList();
     _log.info("Evaluating ${list.length} preconditions");
     try {
       _runningCount++;
-      var results = list.map((p) => p._evaluate());
+      var results = list.map((p) => p._evaluate(ignoreCache: ignoreCache));
       notifyListeners();
       await Future.wait(results);
     } finally {
@@ -139,12 +146,14 @@ class PreconditionsRepository extends ChangeNotifier {
   /// The [registerPrecondition.satisfiedCache] and [registerPrecondition.notSatisfiedCache] can also influence
   /// whether the precondition will be actually run or not.
   ///
-  Future<Precondition> evaluatePreconditionById(String id) async {
+  /// Use [ignoreCache] = true to omit any cached value
+  ///
+  Future<Precondition> evaluatePreconditionById(PreconditionId id, {bool ignoreCache: false}) async {
     var p = _known[id];
     if (p == null) {
       throw Exception("Precondition id = $id is not registered");
     }
-    return evaluatePrecondition(p);
+    return evaluatePrecondition(p, ignoreCache: ignoreCache);
   }
 
   ///
@@ -155,12 +164,14 @@ class PreconditionsRepository extends ChangeNotifier {
   /// The [registerPrecondition.satisfiedCache] and [registerPrecondition.notSatisfiedCache] can also influence
   /// whether the precondition will be actually run or not.
   ///
-  Future<Precondition> evaluatePrecondition(Precondition p) async {
+  /// Use [ignoreCache] = true to omit any cached value
+  ///
+  Future<Precondition> evaluatePrecondition(Precondition p, {bool ignoreCache: false}) async {
     _log.info("Evaluating $p");
     try {
       _runningCount++;
       notifyListeners();
-      await p._evaluate();
+      await p._evaluate(ignoreCache: ignoreCache);
     } finally {
       _runningCount--;
     }
@@ -240,9 +251,9 @@ class Precondition extends ChangeNotifier {
   final StatusBuilder statusBuilder;
 
   /// Identification of this precondition. Supply your own or it will be assigned by the repository.
-  final String id;
+  final PreconditionId id;
 
-  final Iterable<String> dependsOn;
+  final Iterable<PreconditionId> dependsOn;
 
   final PreconditionsRepository _parent;
 
@@ -282,12 +293,12 @@ class Precondition extends ChangeNotifier {
   /// Builds a widget with status description. Uses [statusBuilder] supplied in [PreconditionsRepository.registerPrecondition].
   Widget build(BuildContext context) => statusBuilder(context, status);
 
-  Future<PreconditionStatus> _evaluate() async {
+  Future<PreconditionStatus> _evaluate({bool ignoreCache: false}) async {
     if (_workingOn != null) {
       return await _workingOn!;
     }
     try {
-      _workingOn = _evaluateImpl();
+      _workingOn = _evaluateImpl(ignoreCache: ignoreCache);
       return await _workingOn!;
     } finally {
       _workingOn = null;
@@ -295,8 +306,7 @@ class Precondition extends ChangeNotifier {
   }
 
   @override
-  bool operator ==(Object other) =>
-      identical(this, other) || other is Precondition && runtimeType == other.runtimeType && id == other.id;
+  bool operator ==(Object other) => identical(this, other) || other is Precondition && runtimeType == other.runtimeType && id == other.id;
 
   @override
   int get hashCode => id.hashCode;
@@ -306,15 +316,16 @@ class Precondition extends ChangeNotifier {
     return 'Precondition{#$id, status=$_currentStatus}';
   }
 
-  Future<PreconditionStatus>? _evaluateImpl() async {
-    if (_lastEvaluation != null &&
+  Future<PreconditionStatus>? _evaluateImpl({bool ignoreCache: false}) async {
+    if (!ignoreCache &&
+        _lastEvaluation != null &&
         dependenciesStrategy == DependenciesStrategy.stayInSuccessCache &&
         satisfiedCache.inMicroseconds > 0 &&
         _currentStatus.isSatisfied &&
         _lastEvaluation!.add(satisfiedCache).isAfter(DateTime.now())) return _currentStatus;
 
     if (dependsOn.isNotEmpty) {
-      var ancestors = dependsOn.map((id) => _parent._known[id]).map((p) => p!._evaluate());
+      var ancestors = dependsOn.map((id) => _parent._known[id]).map((p) => p!._evaluate(ignoreCache: ignoreCache));
       var results = await Future.wait(ancestors);
       if (results.any((s) => s.isNotSatisfied)) {
         _currentStatus = PreconditionStatus.unsatisfied();
@@ -323,13 +334,15 @@ class Precondition extends ChangeNotifier {
         return _currentStatus;
       }
     }
-    _log.info("$this - evaluating");
-    if (_lastEvaluation != null &&
+    _log.info("$this - evaluating (ignoreCache=$ignoreCache)");
+    if (!ignoreCache &&
+        _lastEvaluation != null &&
         satisfiedCache.inMicroseconds > 0 &&
         _currentStatus.isSatisfied &&
         _lastEvaluation!.add(satisfiedCache).isAfter(DateTime.now())) return _currentStatus;
 
-    if (_lastEvaluation != null &&
+    if (!ignoreCache &&
+        _lastEvaluation != null &&
         notSatisfiedCache.inMicroseconds > 0 &&
         _currentStatus.isNotSatisfied &&
         _lastEvaluation!.add(notSatisfiedCache).isAfter(DateTime.now())) return _currentStatus;